+++ example/POC_detect_punctuation/README.md
@@ -2,4 +2,3 @@
 
 The files in this folder are used to detect sentence ending punctuation in the transcripts.
 The idea is to have a clear cut in the transcripts by cutting at the end of each whole sentence.
-We are currently detecting the punctuation in the partial transcripts and cutting the audio by the according timestamps.
