+++ example/POC_parallel_transcription/transcribe_one_model.py
@@ -15,8 +15,8 @@ model = WhisperModel(
 )
 
 files = [
-    "scholz_in_kurz.wav",
-    "scholz_in_kurz.wav",
+    "audio.wav",
+    "audio.wav",
 ]
