+++ example/POC_parallel_transcription/transcribe_one_model_multi_gpu.py
@@ -15,9 +15,9 @@ def run_transcription_one_gpu(gpu_index: int):
     )
 
     files = [
-        "scholz_in_kurz.wav",
-        "scholz_in_kurz.wav",
-        "scholz_in_kurz.wav",
+        "audio.wav",
+        "audio.wav",
+        "audio.wav",
     ]
 
     def transcribe_file(file_path, counter):
