+++ example/POC_vosk_transcription/vosk-models/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
