--- example/media-night-frontend/user_interface.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Real-Time Transcription</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            text-align: center;
-            margin-top: 50px;
-        }
-        #transcription {
-            margin-top: 20px;
-            padding: 10px;
-            border: 1px solid #ccc;
-            width: 80%;
-            margin-left: auto;
-            margin-right: auto;
-        }
-        #startButton, #stopButton {
-            padding: 10px 20px;
-            font-size: 16px;
-        }
-    </style>
-</head>
-<body>
-    <h1>Real-Time Transcription</h1>
-    <button id="startButton">Start Microphone</button>
-    <button id="stopButton" style="display: none;">Stop Microphone</button>
-    <div id="transcription">Transcription will appear here...</div>
-
-    <script>
-        let websocket;
-        let mediaRecorder;
-        let audioChunks = [];
-
-        document.getElementById('startButton').addEventListener('click', async () => {
-            document.getElementById('startButton').style.display = 'none';
-            document.getElementById('stopButton').style.display = 'inline';
-
-            websocket = new WebSocket('ws://localhost:8764');
-            websocket.onmessage = (event) => {
-                const data = JSON.parse(event.data);
-                if (data.partial) {
-                    document.getElementById('transcription').innerText = data.partial;
-                } else if (data.result) {
-                    document.getElementById('transcription').innerText = data.text;
-                }
-            };
-
-            const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-            mediaRecorder = new MediaRecorder(stream);
-
-            mediaRecorder.ondataavailable = (event) => {
-                audioChunks.push(event.data);
-                if (mediaRecorder.state === 'inactive') {
-                    const audioBlob = new Blob(audioChunks, { type: 'audio/wav' });
-                    audioChunks = [];
-                    audioBlob.arrayBuffer().then(buffer => websocket.send(buffer));
-                }
-            };
-
-            mediaRecorder.start(1000); // Send audio data in chunks of 1 second
-        });
-
-        document.getElementById('stopButton').addEventListener('click', () => {
-            document.getElementById('startButton').style.display = 'inline';
-            document.getElementById('stopButton').style.display = 'none';
-
-            mediaRecorder.stop();
-            websocket.close();
-        });
-    </script>
-</body>
-</html>
