--- example/vosk/vosk-models/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
