--- infrastructure/README.md
@@ -1,6 +1,0 @@
-# infrastructure
-
-This directory shall be copied to the server running te ASR-API.
-It includes useful files arount the deployment. 
-
-For more information see the docs in the root level `README.md`.
