+++ requirements.txt
@@ -9,6 +9,6 @@ pytest-cov==4.1.0
 websockets==12.0
 SpeechRecognition==3.10.1
 ruff==0.1.15
-numpy<2
+numpy<2 # Required by onnxruntime
 onnxruntime==1.18.0
-onnxruntime-gpu==1.18.0
+onnxruntime-gpu==1.18.0 # Optional, not required for CPU only
